Use PR# as image tag
This will allow us to more easily infer which artifact we should promote
to the `release' repo upon merge. If we use dynamically generated tag names
(such as the epoch time, build id, etc), it's impossible to know which image
was deployed to dev and tested at a later date unless we persist that in-
formation somewhere and retrieve it. For now, we can just scrape the PR#
out of the commit log.